Move all BaseHook usages to version_compat in HTTP
Part of https://github.com/apache/airflow/issues/52676
