+++ providers/http/src/airflow/providers/http/hooks/http.py
@@ -34,11 +34,7 @@ from requests_toolbelt.adapters.socket_options import TCPKeepAliveAdapter
 
 from airflow.exceptions import AirflowException
 from airflow.providers.http.exceptions import HttpErrorException, HttpMethodException
-
-try:
-    from airflow.sdk import BaseHook
-except ImportError:
-    from airflow.hooks.base import BaseHook as BaseHook  # type: ignore
+from airflow.providers.http.version_compat import BaseHook
 
 if TYPE_CHECKING:
     from aiohttp.client_reqrep import ClientResponse
