+++ providers/http/src/airflow/providers/http/operators/http.py
@@ -27,13 +27,8 @@ from requests import Response
 
 from airflow.configuration import conf
 from airflow.exceptions import AirflowException
-
-try:
-    from airflow.sdk import BaseHook
-except ImportError:
-    from airflow.hooks.base import BaseHook  # type: ignore[attr-defined,no-redef]
 from airflow.providers.http.triggers.http import HttpTrigger, serialize_auth_type
-from airflow.providers.http.version_compat import BaseOperator
+from airflow.providers.http.version_compat import BaseHook, BaseOperator
 from airflow.utils.helpers import merge_dicts
 
 if TYPE_CHECKING:
